Changed the percentages on random answer generation.
The code favored generating answers with completely random stems. This worked on the original sample, but because the branch:leaf pairs were all definition: number value. So any given number value could be the answer for several different definitions. After testing with samples where this branch:leaf pair is not so variable, I realized that a lot of the randomization was coming up with choices that very obviously were not correct.

The new random_answer generation has a 96% chance to return a leaf from a random trunk, but with the same branch as the correct answer. There is 2% for completely random trunk and branch, and a 2% chance for a random branch from the correct answer's trunk.

These sliders may eventually be hooked to external input. But for right now, I just need to make quizzes to practice on.